Move CommonMessages to common module to remove duplication
