+++ arktwin/center/src/main/scala/arktwin/center/actors/Atlas.scala
@@ -5,8 +5,8 @@ package arktwin.center.actors
 import arktwin.center.actors.ChartRecorder.ChartRecord
 import arktwin.center.configs.AtlasConfig
 import arktwin.center.util.CenterKamon
-import arktwin.center.util.CommonMessages.Timeout
 import arktwin.common.util.BehaviorsExtensions.*
+import arktwin.common.util.CommonMessages.Timeout
 import arktwin.common.util.MailboxConfig
 import org.apache.pekko.actor.typed.SpawnProtocol.Spawn
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
