+++ arktwin/center/src/main/scala/arktwin/center/actors/Register.scala
@@ -3,8 +3,8 @@
 package arktwin.center.actors
 
 import arktwin.center.services.*
-import arktwin.center.util.CommonMessages.Nop
 import arktwin.common.util.BehaviorsExtensions.*
+import arktwin.common.util.CommonMessages.Nop
 import arktwin.common.util.MailboxConfig
 import org.apache.pekko.actor.typed.SpawnProtocol.Spawn
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
