--- arktwin/center/src/main/scala/arktwin/center/util/CommonMessages.scala
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright 2024-2025 TOYOTA MOTOR CORPORATION
-package arktwin.center.util
-
-object CommonMessages:
-  object Nop
-  object Terminate
-  object Timeout
