+++ arktwin/center/src/test/scala/arktwin/center/actors/AtlasSpec.scala
@@ -6,9 +6,9 @@ import arktwin.center.configs.AtlasConfig
 import arktwin.center.services.ChartAgent
 import arktwin.center.test.ActorTestBase
 import arktwin.center.util.CenterKamon
-import arktwin.center.util.CommonMessages.Terminate
 import arktwin.common.data.*
 import arktwin.common.data.TimestampExtensions.*
+import arktwin.common.util.CommonMessages.Terminate
 import org.apache.pekko.actor.typed.scaladsl.AskPattern.Askable
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
 import org.apache.pekko.actor.typed.{ActorRef, Scheduler}
