+++ arktwin/common/src/main/scala/arktwin/common/util/CommonMessages.scala
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2024-2025 TOYOTA MOTOR CORPORATION
-package arktwin.edge.util
+package arktwin.common.util
 
 object CommonMessages:
   object Nop
+  object Terminate
   object Timeout
