+++ arktwin/edge/src/main/scala/arktwin/edge/actors/adapters/EdgeAgentsPutAdapter.scala
@@ -6,6 +6,7 @@ import arktwin.center.services.ClockBaseExtensions.*
 import arktwin.center.services.{ChartAgent, ClockBase, RegisterAgentUpdated}
 import arktwin.common.data.{MachineTimestamp, TransformEnu}
 import arktwin.common.util.BehaviorsExtensions.*
+import arktwin.common.util.CommonMessages.Timeout
 import arktwin.common.util.{MailboxConfig, VirtualDurationHistogram}
 import arktwin.edge.actors.EdgeConfigurator
 import arktwin.edge.actors.sinks.{Chart, Clock}
@@ -13,7 +14,6 @@ import arktwin.edge.configs.{CoordinateConfig, StaticEdgeConfig}
 import arktwin.edge.connectors.{ChartConnector, RegisterConnector}
 import arktwin.edge.endpoints.EdgeAgentsPut
 import arktwin.edge.endpoints.EdgeAgentsPut.{Request, Response}
-import arktwin.edge.util.CommonMessages.Timeout
 import arktwin.edge.util.ErrorStatus.ServiceUnavailable
 import arktwin.edge.util.{EdgeKamon, ErrorStatus}
 import org.apache.pekko.actor.typed.SpawnProtocol.Spawn
