+++ arktwin/edge/src/main/scala/arktwin/edge/actors/adapters/EdgeNeighborsQueryAdapter.scala
@@ -7,6 +7,7 @@ import arktwin.center.services.{ClockBase, RegisterAgent}
 import arktwin.common.data.TimestampExtensions.*
 import arktwin.common.data.TransformEnuExtensions.*
 import arktwin.common.util.BehaviorsExtensions.*
+import arktwin.common.util.CommonMessages.Timeout
 import arktwin.common.util.{MailboxConfig, VirtualDurationHistogram}
 import arktwin.edge.actors.EdgeConfigurator
 import arktwin.edge.actors.sinks.Chart.CullingAgent
@@ -15,7 +16,6 @@ import arktwin.edge.configs.{CoordinateConfig, StaticEdgeConfig}
 import arktwin.edge.data.Transform
 import arktwin.edge.endpoints.EdgeNeighborsQuery.{Request, Response, ResponseAgent}
 import arktwin.edge.endpoints.{EdgeNeighborsQuery, NeighborChange}
-import arktwin.edge.util.CommonMessages.Timeout
 import arktwin.edge.util.ErrorStatus.ServiceUnavailable
 import arktwin.edge.util.{EdgeKamon, ErrorStatus}
 import org.apache.pekko.actor.typed.SpawnProtocol.Spawn
