+++ arktwin/edge/src/main/scala/arktwin/edge/actors/sinks/Clock.scala
@@ -4,9 +4,9 @@ package arktwin.edge.actors.sinks
 
 import arktwin.center.services.ClockBase
 import arktwin.common.util.BehaviorsExtensions.*
+import arktwin.common.util.CommonMessages.Nop
 import arktwin.common.util.MailboxConfig
 import arktwin.edge.configs.StaticEdgeConfig
-import arktwin.edge.util.CommonMessages.Nop
 import org.apache.pekko.actor.typed.SpawnProtocol.Spawn
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
 import org.apache.pekko.actor.typed.{ActorRef, Behavior}
