+++ arktwin/edge/src/main/scala/arktwin/edge/connectors/ChartConnector.scala
@@ -5,11 +5,11 @@ package arktwin.edge.connectors
 import arktwin.center.services.{ChartAgent, ChartClient, ChartPublishBatch}
 import arktwin.common.data.TimestampExtensions.*
 import arktwin.common.data.{MachineTimestamp, TaggedTimestamp}
+import arktwin.common.util.CommonMessages.Nop
 import arktwin.common.util.GrpcHeaderKey
 import arktwin.common.util.SourceExtensions.*
 import arktwin.edge.actors.sinks.Chart
 import arktwin.edge.configs.StaticEdgeConfig
-import arktwin.edge.util.CommonMessages.Nop
 import arktwin.edge.util.EdgeKamon
 import com.google.protobuf.empty.Empty
 import org.apache.pekko.actor.typed.ActorRef
