+++ arktwin/edge/src/main/scala/arktwin/edge/connectors/RegisterConnector.scala
@@ -3,11 +3,11 @@
 package arktwin.edge.connectors
 
 import arktwin.center.services.{RegisterAgentUpdated, RegisterAgentsPublish, RegisterClient}
+import arktwin.common.util.CommonMessages.Nop
 import arktwin.common.util.GrpcHeaderKey
 import arktwin.common.util.SourceExtensions.*
 import arktwin.edge.actors.sinks.Register
 import arktwin.edge.configs.StaticEdgeConfig
-import arktwin.edge.util.CommonMessages.Nop
 import com.google.protobuf.empty.Empty
 import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.stream.typed.scaladsl.{ActorSink, ActorSource}
